# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/health":{"get":{"operationId":"HealthController_check_v1","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"security":[{}],"summary":"Liveness probe","tags":["Health"]}},"/api/v1/health/ready":{"get":{"operationId":"HealthController_ready_v1","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"security":[{}],"summary":"Readiness probe","tags":["Health"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"400":{"description":"Invalid input data or weak password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Email already registered","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{}],"summary":"Register new user and create profile","tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful, returns access and refresh tokens","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{}],"summary":"Login with email and password","tags":["Auth"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"Tokens refreshed successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenResponseDto"}}}},"401":{"description":"Invalid or expired refresh token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{}],"summary":"Refresh access token","tags":["Auth"]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout_v1","parameters":[],"responses":{"204":{"description":"Logged out"}},"security":[{"bearer":[]}],"summary":"Logout (clear cookies)","tags":["Auth"]}},"/api/v1/auth/me":{"get":{"operationId":"AuthController_getProfile_v1","parameters":[],"responses":{"200":{"description":"Current user profile retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"401":{"description":"Unauthorized access","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearer":[]}],"summary":"Get current user profile","tags":["Auth"]}},"/api/v1/auth/{provider}":{"get":{"operationId":"AuthController_oauth_v1","parameters":[{"name":"provider","required":true,"in":"path","description":"OAuth provider name (google, linkedin)","schema":{"type":"string"}}],"responses":{"302":{"description":"Redirect to provider consent screen"},"400":{"description":"Unsupported provider","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{}],"summary":"Initiate OAuth login with a provider (e.g. google, linkedin)","tags":["Auth"]}},"/api/v1/auth/{provider}/callback":{"get":{"operationId":"AuthController_oauthCallback_v1","parameters":[{"name":"provider","required":true,"in":"path","description":"OAuth provider name (google, linkedin)","schema":{"type":"string"}}],"responses":{"200":{"description":"OAuth authentication successful, returns tokens and user profile","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}},"400":{"description":"Missing email or invalid parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Authentication failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{}],"summary":"OAuth callback endpoint for provider","tags":["Auth"]}},"/api/v1/users/profile":{"get":{"operationId":"UsersController_getProfile_v1","parameters":[],"responses":{"200":{"description":"Own user profile","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDataResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearer":[]}],"summary":"Get current authenticated user profile","tags":["Users"]}},"/api/v1/users":{"post":{"operationId":"UsersController_create_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDataResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Email already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearer":[]}],"summary":"Create a new user (admin only)","tags":["Users"]},"get":{"operationId":"UsersController_findAll_v1","parameters":[{"name":"page","required":false,"in":"query","description":"Page number","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Items per page","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}},{"name":"search","required":false,"in":"query","description":"Search by name or email","schema":{"example":"john","type":"string"}},{"name":"role","required":false,"in":"query","description":"Filter by role","schema":{"example":"USER","type":"string","enum":["USER","ADMIN"]}},{"name":"sort","required":false,"in":"query","description":"Sort field","schema":{"default":"createdAt","example":"updatedAt","type":"string","enum":["createdAt","updatedAt","email","name","role"]}},{"name":"order","required":false,"in":"query","description":"Sort order","schema":{"default":"desc","example":"desc","type":"string","enum":["asc","desc"]}}],"responses":{"200":{"description":"Paginated list of users","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserListResponseDto"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearer":[]}],"summary":"List all users (paginated)","tags":["Users"]}},"/api/v1/users/{id}":{"get":{"operationId":"UsersController_findOne_v1","parameters":[{"name":"id","required":true,"in":"path","description":"User ID (UUID)","schema":{"example":"9dd10e8f-0f3c-4295-befb-6109efd2587f","type":"string"}}],"responses":{"200":{"description":"User found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDataResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearer":[]}],"summary":"Get user by ID (admin only)","tags":["Users"]},"patch":{"operationId":"UsersController_update_v1","parameters":[{"name":"id","required":true,"in":"path","description":"User ID (UUID)","schema":{"example":"9dd10e8f-0f3c-4295-befb-6109efd2587f","type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"User updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDataResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearer":[]}],"summary":"Update user","tags":["Users"]},"delete":{"operationId":"UsersController_remove_v1","parameters":[{"name":"id","required":true,"in":"path","description":"User ID (UUID)","schema":{"example":"9dd10e8f-0f3c-4295-befb-6109efd2587f","type":"string"}}],"responses":{"204":{"description":"User deleted"},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearer":[]}],"summary":"Delete user (admin only)","tags":["Users"]}},"/api/v1/profile":{"get":{"operationId":"ProfileController_getProfile_v1","parameters":[],"responses":{"200":{"description":"Profile retrieved successfully"}},"security":[{"bearer":[]}],"summary":"Get user profile","tags":["profile"]},"patch":{"operationId":"ProfileController_updateProfile_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":"Profile updated successfully"},"400":{"description":"Validation failed"}},"security":[{"bearer":[]}],"summary":"Update user profile","tags":["profile"]}},"/api/v1/reference/fields-of-study":{"get":{"operationId":"ReferenceController_getFieldsOfStudy_v1","parameters":[],"responses":{"200":{"description":"Fields of study retrieved successfully"}},"security":[{}],"summary":"Get all fields of study","tags":["profile"]}},"/api/v1/reference/skills-taxonomy":{"get":{"operationId":"ReferenceController_getSkillsTaxonomy_v1","parameters":[],"responses":{"200":{"description":"Skills taxonomy retrieved successfully"}},"security":[{}],"summary":"Get skills taxonomy grouped by category","tags":["profile"]}},"/api/v1/profile/documents":{"post":{"operationId":"DocumentsController_uploadDocument_v1","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadDocumentDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Upload a document","tags":["profile"]},"get":{"operationId":"DocumentsController_getDocuments_v1","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get all user documents","tags":["profile"]}},"/api/v1/profile/documents/{id}/download":{"get":{"operationId":"DocumentsController_getDownloadUrl_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get download URL for a document","tags":["profile"]}},"/api/v1/profile/documents/{id}":{"delete":{"operationId":"DocumentsController_deleteDocument_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Document deleted successfully"},"404":{"description":"Document not found"}},"security":[{"bearer":[]}],"summary":"Delete a document","tags":["profile"]}},"/api/v1/local-storage/{key}":{"get":{"operationId":"LocalStorageController_downloadFile_v1","parameters":[{"name":"key","required":true,"in":"path","schema":{"type":"string"}},{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{}],"summary":"Download file from local storage","tags":["storage"]}},"/api/v1/profile/educations":{"post":{"operationId":"EducationsController_create_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEducationDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Add a new education record","tags":["profile/educations"]},"get":{"operationId":"EducationsController_findAll_v1","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get all education records for the current user","tags":["profile/educations"]}},"/api/v1/profile/educations/{id}":{"patch":{"operationId":"EducationsController_update_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEducationDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update an education record","tags":["profile/educations"]},"delete":{"operationId":"EducationsController_remove_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete an education record","tags":["profile/educations"]}}},"info":{"title":"Levora API","description":"Levora Intelligent Opportunity Discovery Platform API","version":"1.2","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"X-API-Key":{"type":"apiKey","in":"header","name":"X-API-Key"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com","description":"User email address"},"password":{"type":"string","example":"P@ssw0rd123","minLength":8,"description":"Password must be at least 8 characters long and contain at least one letter and one number"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"}},"required":["email","password"]},"UserProfileDto":{"type":"object","properties":{"fullName":{"type":"object","example":"John Doe"},"isDraft":{"type":"boolean","example":true},"completionPct":{"type":"number","example":0}},"required":["isDraft","completionPct"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string","example":"9dd10e8f-0f3c-4295-befb-6109efd2587f"},"email":{"type":"string","example":"user@example.com"},"firstName":{"type":"object","example":"John"},"lastName":{"type":"object","example":"Doe"},"isEmailVerified":{"type":"boolean","example":false},"isActive":{"type":"boolean","example":true},"lastLoginAt":{"type":"object","example":"2026-08-19T10:00:00.000Z"},"createdAt":{"format":"date-time","type":"string","example":"2026-08-19T10:00:00.000Z"},"userProfile":{"$ref":"#/components/schemas/UserProfileDto"},"roles":{"example":["user"],"type":"array","items":{"type":"string"}}},"required":["id","email","isEmailVerified","isActive","createdAt"]},"ApiErrorItem":{"type":"object","properties":{"field":{"type":"string","example":"email","description":"Field that caused the error"},"code":{"type":"string","example":"USER_EMAIL_DUPLICATE","description":"SCREAMING_SNAKE_CASE error code"},"message":{"type":"string","example":"This email is already registered"}},"required":["code","message"]},"ErrorResponse":{"type":"object","properties":{"success":{"type":"boolean","example":false},"status":{"type":"number","example":404},"message":{"type":"string","example":"User not found"},"data":{"type":"object","nullable":true,"example":null},"meta":{"type":"object","nullable":true,"example":null},"errors":{"type":"array","items":{"$ref":"#/components/schemas/ApiErrorItem"}},"timestamp":{"type":"string","example":"2026-04-22T00:00:00+00:00"}},"required":["success","status","message","data","meta","errors","timestamp"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"P@ssw0rd!"}},"required":["email","password"]},"LoginResponseDto":{"type":"object","properties":{"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"refreshToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"user":{"$ref":"#/components/schemas/UserResponseDto"}},"required":["accessToken","refreshToken","user"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}},"RefreshTokenResponseDto":{"type":"object","properties":{"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}},"required":["accessToken"]},"UserDataResponse":{"type":"object","properties":{"success":{"type":"boolean","example":true},"status":{"type":"number","example":200},"message":{"type":"string","example":"OK"},"data":{"$ref":"#/components/schemas/UserResponseDto"},"meta":{"type":"object","nullable":true},"errors":{"type":"object","nullable":true,"example":null},"timestamp":{"type":"string","example":"2026-04-22T00:00:00+00:00"}},"required":["success","status","message","data","errors","timestamp"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"name":{"type":"string","example":"John Doe"}},"required":["email"]},"PaginationMeta":{"type":"object","properties":{"page":{"type":"number","example":1},"limit":{"type":"number","example":20},"total":{"type":"number","example":248},"totalPages":{"type":"number","example":13},"hasNext":{"type":"boolean","example":true},"hasPrev":{"type":"boolean","example":false}},"required":["page","limit","total","totalPages","hasNext","hasPrev"]},"MetaWithPagination":{"type":"object","properties":{"pagination":{"$ref":"#/components/schemas/PaginationMeta"}},"required":["pagination"]},"UserListResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"status":{"type":"number","example":200},"message":{"type":"string","example":"OK"},"data":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}},"meta":{"$ref":"#/components/schemas/MetaWithPagination"},"errors":{"type":"object","nullable":true,"example":null},"timestamp":{"type":"string","example":"2026-04-22T00:00:00+00:00"}},"required":["success","status","message","data","meta","errors","timestamp"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"name":{"type":"string","example":"John Doe"}}},"SkillDto":{"type":"object","properties":{"skillId":{"type":"string"},"proficiency":{"type":"number"}}},"LanguageDto":{"type":"object","properties":{"languageId":{"type":"string"},"proficiency":{"type":"string"}}},"UpdateProfileDto":{"type":"object","properties":{"fullName":{"type":"string"},"dateOfBirth":{"format":"date-time","type":"string"},"nationality":{"type":"string"},"educationLevel":{"type":"string"},"fieldOfStudy":{"type":"array","items":{"type":"string"}},"currentCountry":{"type":"string"},"currentCity":{"type":"string"},"phone":{"type":"string"},"experienceLevel":{"type":"string"},"hasFinancialNeed":{"type":"boolean"},"careerGoals":{"type":"string"},"profilePhotoUrl":{"type":"string"},"skills":{"type":"array","items":{"$ref":"#/components/schemas/SkillDto"}},"languages":{"type":"array","items":{"$ref":"#/components/schemas/LanguageDto"}}}},"UploadDocumentDto":{"type":"object","properties":{}},"CreateEducationDto":{"type":"object","properties":{"degree":{"type":"string"},"major":{"type":"string"},"institution":{"type":"string"},"graduationYear":{"type":"number","example":2024}},"required":["degree","major","institution"]},"UpdateEducationDto":{"type":"object","properties":{"degree":{"type":"string"},"major":{"type":"string"},"institution":{"type":"string"},"graduationYear":{"type":"number","example":2024},"gpaValue":{"type":"number"},"gpaScale":{"type":"string"}}}}}}